Update kernel args on firstboot-complete
This commits ensures kernel args modifications from /etc/pivot/kernel-args
are being applied on both pivot and firstboot_complete_machineconfig.

OKD requires this so that "mitigations" karg could be removed from cmdline,
as MachineConfig's `kernelArguments` doesn't support removing existing arguments.
